+++ toglacier.go
@@ -287,7 +287,6 @@ func backup(backupPaths []string, backupSecret string, b archive.Builder, e arch
 		return errors.WithStack(err)
 	}
 	backupReport.Durations.Send = time.Now().Sub(timeMark)
-	backupReport.Backup.Paths = backupPaths
 
 	if err := s.Save(backupReport.Backup); err != nil {
 		backupReport.Errors = append(backupReport.Errors, err)